Added validation of incoming data using the Bean Validation mechanism
Added @Email, @NotBlank annotations to the fields of the domain classes, indicating the restrictions and messages that will be displayed in case of validation errors.
Modified controllers: add the @Valid annotation to the data received from the user and 2 required arguments: BindingResult - receives validation messages, Model - serves to transfer data from the controller to the view.
For the design of messages used styles from the CSS framework Bootstrap.