Added user administration panel.
Added a template for editing a user, where username and role fields are available for editing.
A new UserController controller has been created and we have written endpoints in it to display the list of users, the user editing form and the processing of changed user data.
We manage user access rights using the @PreAuthorize ("hasAuthority ('ADMIN')" annotation added to UserController and @EnableGlobalMethodSecurity configuration annotation (prePostEnabled = true) added to WebSecurityConfig.