Added user profile.
The ability to edit the status of the user (NOT_ACTIVE: isEnabled returns false, DELETED: isAccountNonLocked returns false) added to the admin panel.
Created a profile.ftl template and its response part in a UserController controller. In the UserService added the method that handles the change of firstname, lastname, password and email of user.